requirements.txt
requests>=2.25.0
beautifulsoup4>=4.9.0
Pillow>=8.0.0
markdown>=3.3.0 